+++ src/types/Config.tsx
@@ -612,10 +612,6 @@ export const ALL_OVERLAYS: IOverlay[] = [
     role: ROLES.USERMANAGEMENT_VIEW,
   },
   {
-    name: OVERLAYS.TECHUSER,
-    role: ROLES.TECHUSER_VIEW,
-  },
-  {
     name: OVERLAYS.SERVICE_REQUEST,
     role: ROLES.APPSTORE_VIEW_SERVICES,
   },
@@ -634,10 +630,6 @@ export const ALL_OVERLAYS: IOverlay[] = [
   {
     name: OVERLAYS.IDP_CONFIRM,
     role: ROLES.IDP_DELETE,
-  },
-  {
-    name: OVERLAYS.IDP_STATUS,
-    role: ROLES.IDP_DISABLE,
   },
   {
     name: OVERLAYS.ADD_IDP,
@@ -668,20 +660,12 @@ export const ALL_OVERLAYS: IOverlay[] = [
     role: ROLES.IDP_SETUP,
   },
   {
-    name: OVERLAYS.DELETE_IDP,
-    role: ROLES.IDP_DELETE,
-  },
-  {
     name: OVERLAYS.DELETE_MANAGED_IDP,
     role: ROLES.IDP_DELETE,
   },
   {
     name: OVERLAYS.ADDUSERS_IDP,
     role: ROLES.IDP_SETUP,
-  },
-  {
-    name: OVERLAYS.IDP_TEST_RUN,
-    role: ROLES.IDP_ADD,
   },
   {
     name: OVERLAYS.REGISTER_OSP,
@@ -700,9 +684,6 @@ export const ALL_OVERLAYS: IOverlay[] = [
   },
   {
     name: OVERLAYS.CONFIRM_USER_ACTION,
-  },
-  {
-    name: OVERLAYS.SAMPLE_FORM,
   },
   {
     name: OVERLAYS.ADD_SERVICE_PROVIDER,